third_party/WebKit/LayoutTests/webaudio/AudioNode/channel-mode-interp-basic.html

<!doctype html>
<html>
  <head>
    <title>Test Setting of channelCountMode and channelInterpretation</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
    <script src="../resources/audit.js"></script>
  </head>

  <body>
    <script>
      // Fairly arbitrary sample rate and number of frames, except the number of
      // frames should be more than a few render quantums.
      var sampleRate = 16000;
      var renderFrames = 10 * 128;

      var audit = Audit.createTaskRunner();

      audit.define("interp", (task, should) => {
        var context = new OfflineAudioContext(1, renderFrames, sampleRate);
        var node = context.createGain();

        // Set a new interpretation and verify that it changed.
        node.channelInterpretation = "discrete";
        var value = node.channelInterpretation;
        should(value, "node.channelInterpretation").beEqualTo("discrete");
        node.connect(context.destination);

        context.startRendering().then(function (buffer) {
          // After rendering, the value should have been changed.
          should(node.channelInterpretation,
             "After rendering node.channelInterpretation")
            .beEqualTo("discrete");
        }).then(() => task.done());
      });

      audit.define("mode", (task, should) => {
        var context = new OfflineAudioContext(1, renderFrames, sampleRate);
        var node = context.createGain();

        // Set a new mode and verify that it changed.
        node.channelCountMode = "explicit";
        var value = node.channelCountMode;
        should(value, "node.channelCountMode").beEqualTo("explicit");
        node.connect(context.destination);

        context.startRendering().then(function (buffer) {
          // After rendering, the value should have been changed.
          should(node.channelCountMode,
            "After rendering node.channelCountMode")
            .beEqualTo("explicit");
        }).then(() => task.done());
      });

      audit.run();
    </script>
  </body>
</html>
