third_party/WebKit/LayoutTests/platform/win/webaudio/AudioNode/channel-mode-interp-basic-expected.txt

This is a testharness.js-based test.
PASS node.channelInterpretation is equal to "discrete". 
PASS After rendering node.channelInterpretation is equal to "discrete". 
PASS node.channelCountMode is equal to "explicit". 
PASS After rendering node.channelCountMode is equal to "explicit". 
Harness: the test ran to completion.
